requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
python-chess>=1.9.0
tqdm>=4.65.0

# For visualization and plotting
matplotlib>=3.7.0

# Optional for better performance
scipy>=1.10.0
scikit-learn>=1.2.0

# For logging and monitoring (optional)
tensorboard>=2.12.0
wandb>=0.15.0